In maps: Distance value type changed from Integer to Double Reason: This interface can now also be applied to real map data such as Open Street Map data, where node positions are sometimes really near and sometimes very far away from each other.

+++ trunk/src/aima/search/map/Map.java
@@ -20,7 +20,7 @@ public interface Map {
 	 * Returns the travel distance between the two specified locations if
 	 * they are linked by a connection and null otherwise.
 	 */
-	public Integer getDistance(String fromLocation, String toLocation);
+	public Double getDistance(String fromLocation, String toLocation);
 
 	/**
 	 * Returns an array with two integers describing the the position of the
